2-puppet_custom_http_response_header.pp

+++ 0x0F-load_balancer/2-puppet_custom_http_response_header.pp
@@ -1,21 +1,25 @@
 #Add a custom header using puppet
 
 exec {'update':
-  provider => shell,
   command  => 'sudo apt-get -y update',
+  unless   => 'dpkg -s nginx 2>/dev/null',
 }
 
 exec {'nginx':
-  provider => shell,
   command  => 'sudo apt-get -y install nginx',
+  unless   => 'dpkg -s nginx 2>/dev/null',
+  require  => Exec['update'],
 }
 
-exec {'add header':
-  provider => shell,
-  command  => 'sudo sed -i "s/server_name _;/server_name _;\n\tadd_header X-Served-By \$hostname;/" /etc/nginx/sites-available/default',
+exec {'add_header':
+  command   => 'sudo sed -i "s/server_name _;/server_name _;\n\tadd_header X-Served-By \$hostname;/" /etc/nginx/sites-available/default',
+  unless    => 'grep -q "add_header X-Served-By \$hostname;" /etc/nginx/sites-available/default',
+  require   => Exec['nginx'],
+  subscribe => File['/etc/nginx/sites-available/default'],
 }
 
-exec { 'restart service':
-  command  => 'sudo service nginx restart',
-  provider => shell,
-} 
+service { 'nginx':
+  ensure    => running,
+  enable    => true,
+  subscribe => Exec['add_header'],
+}
